move constraint code into own package
