+++ pipeline/frontend/yaml/constraint/constraint.go
@@ -1,4 +1,4 @@
-package yaml
+package constraint
 
 import (
 	"fmt"
